Get it working with caching POST requests
Setting of `$http` and `offline` caches must be done before
`offline.start()` otherwise cached requests in localStorage will not be
processed when the app loads.

Also must pass `{offline: true}` as configuration to the `$http` methods
else they all get passed through because the app is online.

Change to es5-compatible code so it can run and be tested on an iPad.